Give examples for send_message parameters
Within the Microsoft Teams application, if you click on
"Copy URL" for a channel, Teams will URL encode the
channel ID. However, that value needs to be un-decoded
in the parameter. By including an example of the channel ID,
it makes it clearer that the : and @ sign need to be written
as-is. The example for the group-id is added for consistency.